Additional improvements/linting

+++ benchmark-scripts/ebs_large_file_benchmark.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 REMOTE_DIRECTORY=ebs_folder
 LOCAL_DIRECTORY=local_test
 TEST_OUTPUT=test_results.txt
@@ -8,8 +10,7 @@ cleanup() {
     rm -rf "$LOCAL_DIRECTORY/src"
     rm -rf "$LOCAL_DIRECTORY/dst"
 
-    rm -rf "$REMOTE_DIRECTORY/$LOCAL_DIRECTORY"
-    rm -rf "$REMOTE_DIRECTORY/$LOCAL_DIRECTORY"
+    rm -rf ${REMOTE_DIRECTORY:?}/${LOCAL_DIRECTORY:?}
 }
 create_source_directories() {
     mkdir -p $LOCAL_DIRECTORY/src
@@ -43,7 +44,7 @@ attempt_burst() {
 }
 
 clear_dest_remote() {
-    rm -rf $REMOTE_DIRECTORY/$LOCAL_DIRECTOR/dst | tee -a $TEST_OUTPUT
+    rm -rf $REMOTE_DIRECTORY/$LOCAL_DIRECTORY/dst | tee -a $TEST_OUTPUT
 }
 
 clear_dest_local() {
@@ -74,9 +75,6 @@ echo "-- Run EBS Tests --" | tee -a $TEST_OUTPUT
 echo "---------------------------LARGE FILES (WRITE) --------------------------------" | tee -a $TEST_OUTPUT
 i=0
 sum=0
-duration_sum=0
-max=0
-min=1000
 while [[ "${i}" -lt "${REPEATS}" ]]; do
     clear_dest_remote
     clear_cache
@@ -89,13 +87,6 @@ while [[ "${i}" -lt "${REPEATS}" ]]; do
     echo "RUN[${i}] - Time Taken (s): $duration | Speed (Gbps): $speed_gbps | Speed (GiB/s): $speed_gbips"  | tee -a $TEST_OUTPUT
     i=$((i + 1))
     sum=$(echo "scale=5;${sum}+${speed_gbps}" | bc -l | awk '{printf("%.5f",$1)}')
-    if [[ $(echo "$speed_gbps > $max" |bc -l)  ]]; then
-        max=$speed_gbps
-    fi
-    echo "min: $min $speed_gbps"
-    if [ $(echo "$speed_gbps < $min" |bc -l) ]; then
-        min=$speed_gbps
-    fi
     done
 average_speed=$(echo "scale=5;${sum}/${REPEATS}" | bc -l | awk '{printf("%.5f",$1)}')
 #average_speed=$(echo "scale=5;${sum}/${REPEATS}" | bc -l | awk '{printf("%.5f",$1)}')
@@ -108,9 +99,6 @@ rm -rf "$LOCAL_DIRECTORY/src"
 echo "--------------------------LARGE FILES (READ)-------------------------------" | tee -a $TEST_OUTPUT
 i=0
 sum=0
-duration_sum=0
-max=0
-min=1000
 while [[ "${i}" -lt "${REPEATS}" ]]; do
     clear_dest_local
     clear_cache
@@ -123,13 +111,6 @@ while [[ "${i}" -lt "${REPEATS}" ]]; do
     echo "RUN[${i}] - Time Taken (s): $duration | Speed (Gbps): $speed_gbps | Speed (GiB/s): $speed_gbips"  | tee -a $TEST_OUTPUT
     i=$((i + 1))
     sum=$(echo "scale=5;${sum}+${speed_gbps}" | bc -l | awk '{printf("%.5f",$1)}')
-    if [[ $(echo "$speed_gbps > $max" | bc -l)  ]]; then
-        max=$speed_gbps
-    fi
-
-    if [ $(echo "$speed_gbps < $min" | bc -l) ]; then
-        min=$speed_gbps
-    fi
     done
 average_speed=$(echo "scale=5;${sum}/${REPEATS}" | bc -l | awk '{printf("%.5f",$1)}')
 echo "Results - Average (Gbps): $average_speed" | tee -a $TEST_OUTPUT
